Speaker Update (#126)
* Speaker Updates

Adding mounting holes to fit the Dayton Audio DMA80-8 which uses a square flange for a round speaker. Also cutting a chunk out of the lower support to allow a screwdriver or allen wrench to secure that speaker (since you have to secure the bracket first.

* Revert "Speaker Updates"

This reverts commit 5d94b491b89f0aa0a26dc446b1b71abe8a57b39e.

* Adjust File Names

Reverted the original mounting bracket, adding the new one with a new suffix to denote its use.