Cyclotron/Hat Lights (#129)
* Diagram Updates

Replacing the PCB images with diagrams for standard connections. In the case of the wand, some connections are swapped for the bargraph module from Frutto.

* Image Updates

Fixing line/box sizing

* Add label-only images

Restoring label-only boards for reference

* Corrections

Fixed diagram source with v1.1 wand board, shrink other images

* UART Update

Correcting text about the UART connection, with graphics

* Speaker Updates

Adding mounting holes to fit the Dayton Audio DMA80-8 which uses a square flange for a round speaker. Also cutting a chunk out of the lower support to allow a screwdriver or allen wrench to secure that speaker (since you have to secure the bracket first.

* Revert "Speaker Updates"

This reverts commit 5d94b491b89f0aa0a26dc446b1b71abe8a57b39e.

* Adjust File Names

Reverted the original mounting bracket, adding the new one with a new suffix to denote its use.

* Speaker Mounts

Another update with a specific support bracket for the Dayton Audio DMA80 which has a much smaller profile and magnet.

* Major doc updates

Adding tons of images to the doc pages, and updating the Hose and Cyclotron mod guides. Adding the initial doc for the Hat lights.

* Update HATS.md

* Update HOSE.md