Add startup sound for standalone wand
This adds a startup sound to the wand if it is in Afterlife mode and is not being used with a GPstar-equipped Proton Pack. This allows the wand to sound a little more "complete" when first turned on since there is no pack connected to make those sounds instead.

Also fixes a few minor typos.