Update CTS Mix Functionality
This proposed change makes CTS Mix into an independent optional fire mode that can be used alongside the video game or spectral fire modes. In this version, CTS Mix works only when the user is firing a Proton Stream and then holds down the Barrel Wing Button. If they release it, it returns to Proton Stream as before. If they release Intensify, the wand stops firing. When not firing, the Barrel Wing Button continues to be a mode switch as if CTS were not enabled.