Minor Wand changes
This changes the gpstar build code to use digitalRead for the Barrel Wing Button and Barrel Safety Switch as a Mega 2560 has digital buffers on A6 and A7, unlike the Nano. Additionally this change disables both serial read and serial write if b_gpstar_benchtest is enabled, which should prevent any synchronization errors from the pack if a standalone-configured Wand is connected for some reason.