Adjusted STLs for Frutto Bargraph
These updates come courtesy of IdleHandsPrinting, who provided them after a user had issues with the Frutto designs breaking and not sitting correctly in the wand. These updated designs reportedly have better fit and are less fragile. This information was provided alongside the files:

The BLOCK file is cut in two parts and are printed in that orientation. The two pieces are super glued or 2 part clear epoxied together.

The MOUNT would ideally be printed on its side where that "arm" for the screw is on the build plate.

'Detect thin walls' should be on in the slicer and use 0.4 or lower line widths and 0.2mm layer height or smaller.