Attenuator Updates
Clean up and speed up some of the UI code, while adjusting timings for proper sync with the pack. Also introduced a new web API to restart the attenuator/wireless device in the event a re-sync is needed.
Added a longer initial delay for the pack to begin the handshake, ensuring the wand serial connection is established and preferences are obtained from EEPROM.