Standalone wand update
This correctly handles the SYSTEM_YEAR functionality in the Neutrona Wand both when connected to a Proton Pack and when operating in standalone mode. In normal operation checks will rely on the SYSTEM_YEAR variable, which is updated by serial data from a Proton Pack. In standalone mode checks will rely on getNeutronaWandYearMode() instead as this is updated by the Neutrona Wand itself.

Also renames ACTION_EEPROM_MENU to ACTION_LED_EEPROM_MENU for clarity, and updates the EEPROM Operation Manual to clarify the default year functionality difference between System (Proton Pack) and Wand.