Neutrona Wand switch handler overhaul
This changes the Barrel Wing Button and the Barrel Safety Switch over to the ezButton library to correct a bug reported by Ian Gendel where the Barrel Wing Button was not working for menu options which used the rotary encoder.

Additionally this removes all the double-dipping switch debounce code from the wand, since all ezButtons have built-in debouncing and as such there is no reason to run additional debounce timers on them.