Allow wand to define firing sounds
This update will allow the WAND_YEAR_MODE setting in a wand's EEPROM to override a pack's firing sounds so that the firing sounds will be synchronized between the wand and pack. This is done by adding WAND_YEAR_MODE as an i_value component alongside the W_SYNC_NOW serial datagram so that when a wand is first connected it will transmit its setting to the pack. It is also handled in the PACKET_WAND datagram by directly accessing the wandConfig.defaultYearModeWand property so that changing the setting via the EEPROM menu in an already-connected wand or via the attenuator will also update this setting on the pack.