Add year firing config, fix wand sounds
This adds b_year_mode_firing_sounds to the wand configuration to enable or disable the feature of having the firing sounds follow the wand's year mode setting. It also fixes the sound effect playback on the wand side so that it too will obey the wand's year mode setting if this bool is enabled.