Refactor serial handshake
This completely refactors the serial handshake process between the Proton Pack and Neutrona Wand. It makes the actual handshake process more state-driven, increases the number of pings the Pack will wait for before starting a timeout from 1 to 2, and creates a new serial data struct specifically for synchronizing initial settings between the Pack and Wand.