Adding logic for a "lockout" feature on the attenuator
By request by one of our users, I'm adding a means to lock out accidental interactions with the center dial on the attenuator. The keypress combination is a press of the center dial followed immediately by flipping the right toggle. It doesn't matter if it's switched on or off, just that it changed. This will stop any turns of the encoder or pressing of the center post. Unlocking takes the same combination of press + immediate change of state for the right toggle.