WAV Trigger music count fix
This fixes a bug where if the WAV Trigger fails to respond properly with RSP_SYSTEM_INFO after booting (for instance due to an SD card issue), the total number of music tracks would be calculated by the code to be a number in the 65000s due to unsigned integer wraparound as a result of subtracting i_last_effects_track from 0. Instead we will now only try to update the music list if RSP_SYSTEM_INFO has been successfully received. We also send a debug message to warn a user if RSP_SYSTEM_INFO was not received.