Fix incorrect Afterlife bootup
When starting the pack from the wand in Afterlife mode, the pack was incorrectly sending P_ALARM_OFF to the wand which was causing the wand (and only the wand) to play the wand bootup sound erroneously. Additionally the pack will now boot quicker when recovering from an error state in Afterlife mode.