Adding an mDNS responder to the Attenuator
I found that there is a built-in mDNS responder which allows for a super-localized DNS broadcast with a common name for the device. By using the same name as the WiFi access point a user can point their browser to "ProtonPack_NNNN.local" and be automatically directed to the device's web UI. This removes some of the confusion of finding the IP address given to the device, whether it was via the private AP or on an external WiFi network.