Swap stereo channels for 84 bootup
For people who built their pack kits according to the original installation instructions (the majority of users), it was specified that the right channel should be the "main" speaker while the left channel should be the "sub" speaker. In this configuration, the main sound should be in the right channel, not the left. The original file is now 'alt'.