V5.3.3 (#366)
* Fix 84/89 cyclotron overheat

This fixes the erratic behavior from the outer cyclotron when overheating in 84/89 mode by asserting a red colour scheme if we are in one of the two colour-shift firing modes, and then only using the special fade code if we are not overheating.

* Update pack binary

* Fix remote overheat failures

Fixes issues where if pack overheating is triggered by the wireless module, it will fail if either the wand is turned off or no wand is connected.

* Update DIY_WAND.md

Updates several issues with the DIY guide that incorrectly referred to the old Arduino Nano pins when referencing the new Mega2560 Pro Mini board. The new board uses the exact same pinout as the GPStar Wand PCB and should be treated accordingly.

* Add GB1 barrel pop SFX

Thanks to Graham Starke from the HasLab Ghostbusters Modders group we now have a clean recording of the GB1 rooftop barrel pop sound.

* Swap stereo channels for 84 bootup

For people who built their pack kits according to the original installation instructions (the majority of users), it was specified that the right channel should be the "main" speaker while the left channel should be the "sub" speaker. In this configuration, the main sound should be in the right channel, not the left. The original file is now 'alt'.

* Update voltage detection limits

* Implement dynamic standalone mode

This change implements the ability to switch the wand to standalone mode dynamically rather than needing to flash a unique binary. To enable standalone mode, bridge the TX and RX pins on the Neutrona Wand board while powering up.

* Introducing initial code for the single-shot blaster

* Minor fixes

Fixes missing definition for the GBFE pack freeze sound, fixes position of volume reset when switching to standalone mode on the wand, minor typo fix

* Fix sync light behavior

This removes b_sync_light as it was an unnecessary variable, and prevents the top LED from getting stuck in the on state when transitioning to standalone mode.

* Update FE brass pack functions

Disables cyclotron switch panel LEDs if brass pack sound is running (resolves #358) and makes it so user-set spectral colors will be respected in spectral custom mode with brass pack sound running (resolves #359).

* Move pin definitions to defines

This changes the constant variables with lowercase names into preprocessor defines with uppercase names. It also clarifies what a few of the previously ambiguous pins actually do.

* Update SINGLESHOT.md

* Update SINGLESHOT.md

* Change inner cyclotron panel to use an ENUM

* Adjust color coding for static vs. dynamic panel

* Adjust button-smash sounds for GB:FE

* Adjust wand button mash for GB:FE

* Adjust inner LED panel colors, fix typos

This removes unnecessary includes, does proper integer conversion from booleans in Serial.h, and corrects the colors for the year mode selection switch inner cyclotron switch panel.

* Adjust mash timeout, separate compile scripts

* Update compile_attenuator_esp.sh

* Update SINGLESHOT.md

* Additional brass pack fixes

Fixes Frutto LED panel not remaining off while ramping down which resolves #361 and adds a new brass pack shutdown sound from Aces6469

* Brass pack switch panel updates

Updates the brass pack switch panel shutoff functionality to also occur when ribbon cable alarm is active, when overheat is active, and when the button mash freeze lockout is active.

* Adjust freeze-up sound with new shutdown segment, add voices for inner panel options

* Implement more of the freeze-up effect

* Recompile binary

* Add missing sound defs to wand

* Cleanup wand bargraph code

There were a lot of unnecessary calls to the full bargraph reset function and lots of unnecessary explicit multiplier assignments which are better suited to a function dedicated to setting the multiplier.

* Fix improper bargraph speed after overheat

Moved WAND_ACTION_STATUS update to after the bargraph speed reset function so that said reset function has an opportunity to use the correct variables for recovering from overheating.

* Address change to mode for ezButton library

* More work on the pack freeze-up for GBFE

* Adjusting serial comms to attenuator (for stock wand)

* Return smoke prefs when no wand is connected

* Adjustments for power detection and display

* Rebuild binaries

* Text adjustments for UI

* Update PowerMeter.h

* Change from Amps to Watts for measurements

* Update PowerMeter.h

* Update PowerMeter.h

* Minor text adjustments for UI

* Removed some debugs, changed color values for more obvious effect

* stl updates

* Wand overheating fixes, optimizations

Fixes wand idle hum overlap when recovering from overheat, fixes overlapping pack startup sounds when recovering from overheat, makes it so wand bootup sounds will only be sent if the pack is not already playing its own bootup sound, fixes wand function names to more accurately describe what they do, cleaned up code relating to overheat an AL/FE startup ramps, removed legacy wand variables from single shot blaster.

* updated top dial no numbers fix

* Minor timing tweaks

* Move overheat into dedicated function

This moves the overheating initiator on the pack into a dedicated function rather than having the functionality duplicated across three separate call locations. Likewise it creates a dedicated function for the quick vent start for future extensibility.

* Fix pack buttons in web interface

This fixes the pack on/off buttons inadvertently ignoring pack state, and fixes the manual vent button not working unless a wand is connected (which is no longer necessary).

* Improve 84/89 cyclotron fade effect

Extends the fade effect for the 84/89 cyclotron to fade out after one additional movement, which both makes the effect look less jerky and more closely represents how the animation looked in the video game.

* Add power notes for Singleshot

* Fix 84/89 cyclotron overheat animation

This resolves #354 by fixing the out-of-sequence cyclotron LEDs when overheat is in progress in 84/89 year mode.

* Rebuild with new Switch library

* Fix 84/89 cyclotron in multicolor modes

This fixes the seizure-inducing outer cyclotron LEDs in Holiday and Spectral firing modes for good. While there is room for improvement that can come with v6 when the LED code gets overhauled.

* Fix powercell sound playing when button mashing

This fixes the Afterlife/Frozen Empire powercell SFX still playing during button mash lockout.

* Additional fix to AL/FE powercell SFX

Additional conditional checks added to Afterlife/Frozen Empire powercell sound.

* Fix potential button mash flag issue

Makes sure that if we start a button mash sequence on the pack and the pack thinks the wand is firing, we make sure that we reset all the firing-related flags and sounds.

* fixed powercell sound effect but

* Additional button mash cleanup

Prevents noisy overlaps of wand restart sounds in 84/89, makes certain that the button mash restart will always occur even when extra wand sounds are disabled, and cleans up some powercell and button mash checks on the pack.

* Recompile wand binaries

* Whitespace fix

Fixes missing whitespace before braces

* fixed quick volume adjustment bug from the wand

---------

Co-authored-by: nomakewan <nomakewan@users.noreply.github.com>
Co-authored-by: Michael Rajotte <michael.rajotte@gmail.com>