Update POST sequencing
This change allows the pack to tell a wand attempting to synchronize during the power on self test sequence that it is performing a POST and so should not allow the user to access normal wand functions until it is complete. This prevents, for example, someone entering the EEPROM menu before the POST sequence has been completed. It also prevents switches on the pack from having an effect until the POST sequence completes for the same reason.