Refactor the Attenuator code to use VSCode with PlatformIO
This was an inevitable change based on the future work, but was forced to be done soon as we can use distinct versions of libraries for builds using this approach. The older Nano build may need to stay locked to the FastLED v3.6.0 library in order to run correctly, so this move allows for that to be accomplished.