Audio subsystem overhaul
This updates the audio subsystem in all devices to properly calculate track volume (which should always be -70 to 0) and master volume (which changes depending on device and which sound board is in use). It also cleans up some of the audio-related code making it more concise and readable. It also unifies the wand mute request into a single function.