Update N-Filter LED functionality
This updates the N-Filter LEDs so that they only flash when overheating, when the ribbon cable alarm is active, or when continuous firing smoke is enabled for the current wand power level. In all other cases, flashing is disabled (which allows users to thus disable flashing the N-Filter LEDs by turning off continuous fire smoke).