RGB Vent Light optimization
This optimizes the vent light control to remove the boolean since both the addressable and non-addressable LEDs are updated in parallel, and thus we can use the boolean state of the existing addressable LED memory object to tell whether the LEDs are on or off. It also corrects the colors of the vent light and top LED depending on year (84/89 have a bright white vent and warm white top LED; AL/FE have a warm white vent and bright white top LED), and makes it so only the vent changes colors with firing mode changes.