Vent light off adjustment
Adjusts the code for turning the addressable vent light off so that if requested to turn off and it is on, it will turn off even if the control timer has not yet fired. This prevents issues where an 'off' command may only be fired once and gets 'missed' due to the timer.