Update EEPROM functionality
This cleans up some of the EEPROM functionality to be more efficient and correctly use a constant variable for the starting address in the EEPROM object, and corrects the EEPROM brightness values on the pack to only allow a maximum of 100. It also changes the write routine for clearing the EEPROM to use 0xFF which is the correct default value for ATMEGA EEPROM objects.