Fix wand not obeying pack music status
This fixes issue where wand volume controls for music were only working with a standalone wand by making sure we transmit the current music playback state to the wand if it changes for any reason. It also makes sure that if music is stopped by an ESP32 while the system is off, the power off reminder will correctly activate.