V6.0.1 (#441)
* Fix 0 values for Hue/Saturation

Fixes issue where a value of 0 for hue or saturation on the custom spectral settings would instead send a default value to the pack or wand.

* Rebuild binaries for ESP32

* Backporting web UI fixes

* Fix FastLED crash

Fixes core abort caused by FastLED animation task smashing the stack by doubling the stack size for that task.

* Additional ESP32 fixes

Adds the stack size fix to the BeltGizmo and StreamEffects devices as well.

* Audio init update

This updates how the audio subsystem initialization step works to greatly improve boot times.

* Close serial3 after failed sync

Since we do not attempt audio sync after failing once, best practice is to close the unused serial port so that stray interference cannot trigger serial interrupts.

* Minor adjustments

Makes sure that the minimum percentage brightness able to be read from the EEPROM is 20%, makes the default timeout for the GPStar Audio 1s, corrects minimum power value for GPStar Lite, corrects GPStar Lite wand refresh rate to precisely 20ms, adds firing stop threshold.

* Implement an EEPROM reset firmware

* Rebuild all binaries

* Change to use EEPROM.update()

* Add infinite loop to halt program

* Update gpstarFirmwareFlasher.exe

* Reinforce EEPROM checks

This reinforces the EEPROM checks across all three devices to make sure that their saved settings are all of sane values before loading them into runtime variables.

* Update EEPROM functionality

This cleans up some of the EEPROM functionality to be more efficient and correctly use a constant variable for the starting address in the EEPROM object, and corrects the EEPROM brightness values on the pack to only allow a maximum of 100. It also changes the write routine for clearing the EEPROM to use 0xFF which is the correct default value for ATMEGA EEPROM objects.

* Fix wand not obeying pack music status

This fixes issue where wand volume controls for music were only working with a standalone wand by making sure we transmit the current music playback state to the wand if it changes for any reason. It also makes sure that if music is stopped by an ESP32 while the system is off, the power off reminder will correctly activate.

* Recompile binaries

* Fix stuck pixel, clean up code

Fixes issue where there can be a stuck pixel with a 36-LED outer cyclotron if the animation is on the last 4 LEDs when the user changes firing modes, cleans up some array calls to reduce dynamic memory allocation, and renames some legacy functions and variables that unnecessarily call themselves "2021".

* Fix 84/89 color reset

84/89 does use parallel arrays, so put in system year check to switch paradigm depending on which year cyclotron is currently in use.

* Fix AL/FE cyclotron fades

Fixes additional non-parallel array comparison in outer cyclotron code relating to how the cyclotron fades out.

* Revert cyclotronLidLedsOff()

Reverts the cyclotronLidLedsOff() function to an older version to correct an issue where the entire array may not be shut off when a shutdown is commanded.

* Revert slime, spectral changes

Reverts array traversal for slime effect and spectral light activation to previous paradigm due to inadvertent shortening of the for loops.

* Whitespace fix

Removes unnecessary whitespace in cpp file conditional statements.

* Rebuild binaries

---------

Co-authored-by: nomakewan <nomakewan@users.noreply.github.com>
Co-authored-by: Michael Rajotte <35165105+gpstar81@users.noreply.github.com>