Add GPIO reassignment function
This adds the PIN_FUNC_SELECT function in order to set GPIO39~GPIO44 to GPIO in the IOMUX register, overriding their defaults as JTAG and UART0 pins. It also corrects the serial initialization in the Single Shot Blaster, and adds a command to detach UART0 from Serial0 on the pack as this UART will be reassigned to Serial2.