Whitespace for readability, regroup preprocessors
When using the architecture preprocessors we should probably try to group those around logic such that we don't run if/else statements just to discard their CPU-specific actions. This avoid unnecessary CPU ops--unless such statements need to cause an exit of the statement or an outer loop.

Case statements would still be used as-is because we don't want to end up in a default case by skipping a possible value.