Add stream switcher guardrails
Adds check to stream switcher in webUI to only enable it if the pack is in Mode Super Hero and the wand has VG modes enabled or is disconnected, and adds necessary call to request wand settings on connection for pack if no attenuator is connected. Also defaults the stream switcher to proton mode in case an invalid selection is returned.