Implement barrel switch polarity
Implements ability to select the polarity of the barrel safety switch in case someone is using a non-Hasbro setup for their wand barrel pop mechanism, or if they want to disable the barrel position detection functionality entirely. This change also reintroduces the ability for the wand to toggle its own auto vent light intensity setting via the LED EEPROM menu, and moves the pack's VG color toggles to the LED EEPROM menu.